feat 升级 mp spring boot版本, 优化数据范围权限插件, 修改entity使用数据库关键字导致jsqlparser无法解析的问题

+++ bootx-commons/common-mybatis-plus/src/main/java/cn/bootx/common/mybatisplus/impl/BaseManager.java
@@ -381,7 +381,7 @@ public class BaseManager<M extends BaseMapper<T>, T>{
      * @param fieldValue 字段数据
      * @return 数量
      */
-    public Integer countByField(SFunction<T, ?> field, Object fieldValue){
+    public Long countByField(SFunction<T, ?> field, Object fieldValue){
         return lambdaQuery().eq(field,fieldValue).count();
     }
